fix msg_sent serialization

+++ tg_speaking_from_heart/src/database/database.go
@@ -9,7 +9,7 @@ type DB interface {
 	IsSubscriber(subs Subscriber) bool
 	RandomSubscriber() (Subscriber, error)
 	IncMsgSent()
-	MsgSentCount() int
+	MsgSentCount() uint64
 	SubsCount() int
 	AsJson() string
 }
